Cleaning up all MBR tests.
